requirements.txt
fastapi
uvicorn
pydantic
python-multipart
PyMuPDF
google-generativeai
